- refactoring: export, import - fixed error on empty value - readme: added api
