+++ modules/config/default.js
@@ -1,5 +1,5 @@
 import en_US from 'antd/lib/locale-provider/en_US';
-import * as Widgets from '../components/widgets/index.js';
+import * as Widgets from '../components/widgets';
 import React from "react";
 const {
   FieldSelect,
