+++ modules/import/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+export * from './tree';
