+++ modules/index.js
@@ -1,10 +1,13 @@
 export { default as Query } from './components/Query';
 export { default as Builder } from './components/Builder';
 
+import * as Export from './export';
+import * as Import from './import';
 import * as Widgets from './components/widgets/index';
 import * as Operators from './components/operators';
-import * as Utils from './utils';
-export {Widgets, Operators, Utils};
+import * as BasicUtils from './utils';
+const Utils = {...BasicUtils, ...Export, ...Import};
+export {Widgets, Operators, Utils, Export, Import};
 
 export {default as BasicConfig} from './config/basic';
