+++ modules/utils/index.js
@@ -1,14 +1,3 @@
 'use strict';
-export {queryBuilderFormat} from './queryBuilderFormat'
-export {jsonLogicFormat} from './jsonLogicFormat'
-export {mongodbFormat} from './mongodbFormat'
-export {sqlFormat} from './sqlFormat'
-export {queryString} from './queryString'
-export {getTree, loadTree, checkTree} from './treeUtils'
 export {validateTree} from './validation'
 export {default as uuid} from './uuid';
-export {
-  getFieldConfig, getFieldRawConfig, getFieldPath, getFieldPathLabels, getValueLabel, extendConfig, 
-  getFieldWidgetConfig, getOperatorConfig, getWidgetsForFieldOp, getWidgetForFieldOp, getValueSourcesForFieldOp
-} from './configUtils';
-
